Color_Extractor/requirements.txt
streamlit
opencv-python-headless
scikit-learn
pandas
Pillow
numpy